Fix date parsing on client for MySQL-like-formatted dates
For some reason (probably due to nightly update of events) I can
not reproduce, sometimes dates are not returned as iso strings but
in MySQL format. This breaks new Date(s) on iOS.

To reproduce the issue:

if (key === 'datetime') {
    // 2015-08-14T18:30:00.000Z
    return value.replace('T', ' ').replace('.000Z', '');
}

This fix tries to parse the date from MySQL format as a fallback.